Se crea el custom hook useFetchGifs con la logica para extraer las imagenes
