+++ src/components/GifGrid.jsx
@@ -1,18 +1,11 @@
-import { useEffect, useState } from 'react';
-import { getGifs } from '../helpers/getGifs';
 import { GifItem } from './GifItem';
+import { useFetchGifs } from '../hooks/useFetchGifs';
 
 export const GifGrid = ({ category }) => {
-   const [images, setImages] = useState([]);
+   const { images, isLoading } = useFetchGifs(category);
 
-   const getImages = async () => {
-      const newImages = await getGifs(category);
-      setImages(newImages);
-   };
-
-   useEffect(() => {
-      getImages();
-   }, []);
+   // console.log(images, isLoading);
+   // console.log({ images, isLoading });
 
    return (
       <>
