+++ src/components/GifItem.jsx
@@ -1,5 +1,7 @@
 export const GifItem = ({ title, url }) => {
-   return <div className='card'>
-    <img src={url} alt={title} />
-   </div>;
+   return (
+      <div className='card'>
+         <img src={url} alt={title} />
+      </div>
+   );
 };
